requirements.txt
langchain
langgraph
langchain-community
langchain-openai
langsmith
faiss-cpu
langchain-text-splitters
python-dotenv
beautifulsoup4
